cleaning up implementation
