+++ src/index.js
@@ -1,0 +1,2 @@
+module.exports = require('./maker.js');
+module.exports.generateIcons = require('./generator.js');
